Use normal edit text for the Amount field instead of the NumberPicker

+++ app/src/main/java/za/co/inflationcalc/model/Result.java
@@ -4,13 +4,13 @@ package za.co.inflationcalc.model;
  * Created by Laurie on 9/14/2015.
  */
 public class Result {
-    private int resultValue;
+    private double resultValue;
 
-    public Result(int resultValue) {
+    public Result(double resultValue) {
         this.resultValue = resultValue;
     }
 
-    public int getResultValue() {
+    public double getResultValue() {
         return resultValue;
     }
 }
